Renamed model files to match the model name
